Set submodule commits to match dependencies on risc0-ethereum (#109)
* remove dependency on murky

* update forge std and openzeppelin versions

* point risc0-ethereum submodule to release-1.0 branch

* switch cargo refs to main

* fix up build issues

* update notes in the README about proving with Groth16

* point to the pending release-1.0 branch

* move TxSender directly into publisher

* remove unused submodule
