+++ README.md
@@ -1,10 +1,5 @@
 # RISC Zero Foundry Template
 
-> **Note: This software is not production ready. Do not use in production.**
-
-> Note: This template was recently updated to work without the relay.
-> If you'd like the relay version, head over to the [relay branch](https://github.com/risc0/bonsai-foundry-template/tree/relay)
-
 Starter template for writing an application using [RISC Zero] and Ethereum.
 
 This repository implements an application on Ethereum utilizing RISC Zero as a [coprocessor] to the smart contract application.
@@ -98,19 +93,27 @@ Your new project consists of:
   RISC0_DEV_MODE=true forge test -vvv 
   ```
 
+- Run the same tests, with the full zkVM prover rather than dev-mode, by setting `RISC0_DEV_MODE=false`.
+
+  ```sh
+  RISC0_DEV_MODE=false forge test -vvv
+  ```
+
+  Producing the [Groth16 SNARK proofs][Groth16] for this test requires running on an x86 machine with [Docker] installed, or using [Bonsai](#configuring-bonsai).
+
 ## Develop Your Application
 
 To build your application using the RISC Zero Foundry Template, you’ll need to make changes in three main areas:
 
--	***Guest Code***: Write the code you want proven in the [methods/guest](./methods/guest/) folder. This code runs off-chain within the RISC Zero zkVM and performs the actual computations. For example, the provided template includes a computation to check if a given number is even and generate a proof of this computation.
--	***Smart Contracts***: Write the on-chain part of your project in the [contracts](./contracts/) folder. The smart contract verifies zkVM proofs and updates the blockchain state based on the results of off-chain computations. For instance, in the [EvenNumber](./contracts/EvenNumber.sol) example, the smart contract verifies a proof that a number is even and stores that number on-chain if the proof is valid.
--	***Publisher Application***: Adjust the publisher example in the [apps](./apps/) folder. The publisher application bridges off-chain computation with on-chain verification by submitting proof requests, receiving proofs, and publishing them to the smart contract on Ethereum.
+- ***Guest Code***: Write the code you want proven in the [methods/guest](./methods/guest/) folder. This code runs off-chain within the RISC Zero zkVM and performs the actual computations. For example, the provided template includes a computation to check if a given number is even and generate a proof of this computation.
+- ***Smart Contracts***: Write the on-chain part of your project in the [contracts](./contracts/) folder. The smart contract verifies zkVM proofs and updates the blockchain state based on the results of off-chain computations. For instance, in the [EvenNumber](./contracts/EvenNumber.sol) example, the smart contract verifies a proof that a number is even and stores that number on-chain if the proof is valid.
+- ***Publisher Application***: Adjust the publisher example in the [apps](./apps/) folder. The publisher application bridges off-chain computation with on-chain verification by submitting proof requests, receiving proofs, and publishing them to the smart contract on Ethereum.
 
 ### Configuring Bonsai
 
 ***Note:*** *To request an API key [complete the form here](https://bonsai.xyz/apply).*
 
-With the Bonsai proving service, you can produce a [Groth16 SNARK proof] that is verifiable on-chain.
+With the Bonsai proving service, you can produce a [Groth16 SNARK proof][Groth16] that is verifiable on-chain.
 You can get started by setting the following environment variables with your API key and associated URL.
 
 ```bash
@@ -123,8 +126,6 @@ Now if you run `forge test` with `RISC0_DEV_MODE=false`, the test will run as be
 ```bash
 RISC0_DEV_MODE=false forge test -vvv
 ```
-
-> You can also generate a[Groth16 SNARK proof] locally, assuming you have a machine with an x86 architecture and [Docker] installed. In this case do not export Bonsai related env variables.
 
 ### Deterministic Builds
 
@@ -175,7 +176,7 @@ Below are the primary files in the project directory
 [Bonsai]: https://dev.bonsai.xyz/
 [Foundry]: https://getfoundry.sh/
 [Get Docker]: https://docs.docker.com/get-docker/
-[Groth16 SNARK proof]: https://www.risczero.com/news/on-chain-verification
+[Groth16]: https://www.risczero.com/news/on-chain-verification
 [RISC Zero Verifier]: https://github.com/risc0/risc0/blob/release-0.21/bonsai/ethereum/contracts/IRiscZeroVerifier.sol
 [RISC Zero installation]: https://dev.risczero.com/api/zkvm/install
 [RISC Zero zkVM]: https://dev.risczero.com/zkvm
