+++ apps/README.md
@@ -42,14 +42,5 @@ Options:
           Print version
 ```
 
-## Library
-
-We provide a small rust [library] containing utility functions to help with sending off-chain proof requests to the Bonsai proving service and publish the received proofs directly to a deployed app contract on Ethereum.
-
-As we continue to improve the [risc0-zkvm] and [bonsai-sdk] crates, we will absorb some of the functionality provided here into those crates.
-
 [publisher]: ./src/bin/publisher.rs
 [Bonsai]: https://dev.bonsai.xyz/
-[library]: ./src/lib.rs
-[risc0-zkvm]: https://docs.rs/risc0-zkvm/latest/risc0_zkvm/
-[bonsai-sdk]: https://docs.rs/bonsai-sdk/latest/bonsai_sdk/
