+++ apps/src/bin/publisher.rs
@@ -19,16 +19,57 @@
 use alloy_primitives::U256;
 use alloy_sol_types::{sol, SolInterface, SolValue};
 use anyhow::{Context, Result};
-use apps::TxSender;
 use clap::Parser;
+use ethers::prelude::*;
 use methods::IS_EVEN_ELF;
-use risc0_ethereum_contracts::groth16::Seal;
+use risc0_ethereum_contracts::groth16;
 use risc0_zkvm::{default_prover, ExecutorEnv, ProverOpts, VerifierContext};
 
 // `IEvenNumber` interface automatically generated via the alloy `sol!` macro.
 sol! {
     interface IEvenNumber {
         function set(uint256 x, bytes calldata seal);
+    }
+}
+
+/// Wrapper of a `SignerMiddleware` client to send transactions to the given
+/// contract's `Address`.
+pub struct TxSender {
+    chain_id: u64,
+    client: SignerMiddleware<Provider<Http>, Wallet<k256::ecdsa::SigningKey>>,
+    contract: Address,
+}
+
+impl TxSender {
+    /// Creates a new `TxSender`.
+    pub fn new(chain_id: u64, rpc_url: &str, private_key: &str, contract: &str) -> Result<Self> {
+        let provider = Provider::<Http>::try_from(rpc_url)?;
+        let wallet: LocalWallet = private_key.parse::<LocalWallet>()?.with_chain_id(chain_id);
+        let client = SignerMiddleware::new(provider.clone(), wallet.clone());
+        let contract = contract.parse::<Address>()?;
+
+        Ok(TxSender {
+            chain_id,
+            client,
+            contract,
+        })
+    }
+
+    /// Send a transaction with the given calldata.
+    pub async fn send(&self, calldata: Vec<u8>) -> Result<Option<TransactionReceipt>> {
+        let tx = TransactionRequest::new()
+            .chain_id(self.chain_id)
+            .to(self.contract)
+            .from(self.client.address())
+            .data(calldata);
+
+        log::info!("Transaction request: {:?}", &tx);
+
+        let tx = self.client.send_transaction(tx, None).await?.await?;
+
+        log::info!("Transaction receipt: {:?}", &tx);
+
+        Ok(tx)
     }
 }
 
@@ -81,12 +122,12 @@ fn main() -> Result<()> {
             env,
             &VerifierContext::default(),
             IS_EVEN_ELF,
-            &ProverOpts::compact(),
+            &ProverOpts::groth16(),
         )?
         .receipt;
 
     // Encode the seal with the selector.
-    let seal = Seal::encode(receipt.inner.compact()?.seal.clone())?;
+    let seal = groth16::encode(receipt.inner.groth16()?.seal.clone())?;
 
     // Extract the journal from the receipt.
     let journal = receipt.journal.bytes.clone();
