+++ methods/guest/Cargo.toml
@@ -12,7 +12,7 @@ path = "src/bin/is_even.rs"
 [dependencies]
 alloy-primitives = { version = "0.6", default-features = false, features = ["rlp", "serde", "std"] }
 alloy-sol-types = { version = "0.6" }
-risc0-zkvm = { git = "https://github.com/risc0/risc0", branch = "main", default-features = false, features = ['std'] }
+risc0-zkvm = { version = "1.0.0-rc.6", default-features = false, features = ['std'] }
 
 [profile.release]
 lto = "thin"
