+++ remappings.txt
@@ -1,5 +1,3 @@
-murky/=lib/murky/
-murky_differential_testing/=lib/murky/differential_testing/
 forge-std/=lib/forge-std/src/
 openzeppelin/=lib/openzeppelin-contracts/
 risc0/=lib/risc0-ethereum/contracts/src/
